Flutter 💙
- Less data for calls
- Code refactored

+++ flutter/whatsapp/lib/app/widgets/cached_network_image_widget.dart
@@ -45,7 +45,7 @@ class CachedNetworkImageWidget extends StatelessWidget {
                 fit: fit,
                 progressIndicatorBuilder: (_, __, progress) {
                   NetworkListeners.downloadFileSizeStream
-                      .add(progress.downloaded);
+                      .add(progress.downloaded.toDouble());
                   return placeholder;
                 },
                 errorWidget: (_, __, dynamic ___) {
